gltf: Move node-related functionality to node.cpp
